Teacher CRUD

+++ DayCare/src/edu/neu/csye6200/api/TeacherApi.java
@@ -22,7 +22,7 @@ public interface TeacherApi {
 
     void deleteTeacher(Teacher teacher);
 
-    void deleteTeacher(int teacherId);
+    void deleteTeacher(long teacherId);
 
     double getRating(int teacherId);
 }
