+++ DayCare/src/edu/neu/csye6200/api/concrete/ConcreteTeacherApi.java
@@ -50,7 +50,7 @@ public class ConcreteTeacherApi implements TeacherApi {
     }
 
     @Override
-    public void deleteTeacher(int teacherId) {
+    public void deleteTeacher(long teacherId) {
         TeacherDao.deleteTeacherDao(teacherId);
     }
